+++ src/Register-AudioRebindTask.ps1
@@ -1,4 +1,4 @@
-<#
+﻿<#
 .SYNOPSIS
   Register an elevated Task Scheduler task that runs AudioRebind on resume.
 
@@ -35,7 +35,7 @@ trap {
     $msg = [string]$_.Exception.Message
     Write-Output $msg
     if ($env:AUDIOREBIND_SETUP_CAPTURE -ne '1') {
-        Show-AudioRebindSetupFailure -Kind (Get-AudioRebindSetupFailureKind -Text $msg) -Detail $msg
+        Show-AudioRebindSetupFailure -Kind (Get-AudioRebindSetupFailureKind -Text $msg)
     }
     exit 1
 }